refactor(console,phrases): move permissions and roles to step 2 in org guide
