+++ packages/console/src/pages/Organizations/Guide/Introduction/index.module.scss
