+++ packages/console/src/pages/Organizations/Guide/const.ts
@@ -1,5 +1,5 @@
 export const steps = Object.freeze({
-  createPermissions: 'create-permissions',
-  createRoles: 'create-roles',
-  createOrganization: 'create-organization',
+  introduction: 'introduction',
+  permissionsAndRoles: 'permissions-and-roles',
+  organizationInfo: 'organization-info',
 });
